+++ js/data.js
@@ -1,0 +1,45 @@
+const POST_COUNT = 25;
+const POST_COMMENT_AVATAR_COUNT = 6;
+
+const LIKES_COUNT = {
+  min: 15,
+  max: 200
+};
+const COMMENT_COUNT = {
+  min: 0,
+  max: 30
+};
+const POST_DESCRIPTION = [
+  'Достойный сын своего народа',
+  'Звать меня бесполезно, лучше буду спокойно отдыхать',
+  'Главный ночной тыгыдык',
+  'Если бы вы знали как мне пофиг на всё, вы бы расплакались',
+  'Просто дайте мне вкусняшку',
+  'Если вы видели плохих котов, то забудьте'
+];
+const POST_COMMENT_MESSAGE = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+const POST_COMMENT_NAME = [
+  'Барсик',
+  'Пушок',
+  'Снежок',
+  'Лапусик',
+  'Паша',
+  'Кнопа'
+];
+
+export {
+  POST_COUNT,
+  LIKES_COUNT,
+  COMMENT_COUNT,
+  POST_DESCRIPTION,
+  POST_COMMENT_MESSAGE,
+  POST_COMMENT_NAME,
+  POST_COMMENT_AVATAR_COUNT
+};
